Refactor ProductPage and Header components; add Features and Faq sections, and update routes
